+++ ui/js/websocket/WebSocketProvider.js
